cahnged rate to float in endpoint

+++ Server/endpoints/favorTransaction.js
@@ -62,9 +62,9 @@ exports.favorTransactionTypeDefs = gql`
                         handleFavorTransactionRequest(favorTransactionId: String!, comment: String): String
                         cancelFavorTransaction(favorTransactionId: String!): String
                         executorNotifyDoneFavor(favorTransactionId: String!): String
-                        ownerNotifyJestaHasBeenDone(favorTransactionId: String!, rate: Int, handlerComment: String): String
+                        ownerNotifyJestaHasBeenDone(favorTransactionId: String!, rate: Float, handlerComment: String): String
                         userChangeJestaTransactionToClosed(favorTransactionId: String!): String
-                        rateTransactionWithOptionalComment(favorTransactionId: String!, rate: Int, handlerComment: String): String
+                        rateTransactionWithOptionalComment(favorTransactionId: String!, rate: Float, handlerComment: String): String
                     }
                     `;
